Move _applyControl method to Player base class
