+++ src/localplayer.h
@@ -20,10 +20,8 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 #pragma once
 
 #include "player.h"
-#include "environment.h"
 #include "constants.h"
 #include "settings.h"
-#include "controllog.h"
 #include <list>
 
 class Client;
@@ -53,14 +51,6 @@ public:
 	// doesn't support health points
 	u16 hp = 0;
 	bool is_attached = false;
-
-	float physics_override_speed = 1.0f;
-	float physics_override_jump = 1.0f;
-	float physics_override_gravity = 1.0f;
-	bool physics_override_sneak = true;
-	bool physics_override_sneak_glitch = false;
-	// Temporary option for old move code
-	bool physics_override_new_move = true;
 
 	v3f overridePosition;
 
@@ -115,19 +105,6 @@ public:
 
 	v3s16 getLightPosition() const;
 
-	void setYaw(f32 yaw) { m_yaw = yaw; }
-	f32 getYaw() const { return m_yaw; }
-
-	void setPitch(f32 pitch) { m_pitch = pitch; }
-	f32 getPitch() const { return m_pitch; }
-
-	inline void setPosition(const v3f &position)
-	{
-		m_position = position;
-		m_sneak_node_exists = false;
-	}
-
-	v3f getPosition() const { return m_position; }
 	v3f getEyePosition() const { return m_position + getEyeOffset(); }
 	v3f getEyeOffset() const;
 	void setEyeHeight(float eye_height) { m_eye_height = eye_height; }
@@ -145,11 +122,8 @@ protected:
 	virtual void triggerJumpEvent();
 
 private:
-	void accelerateHorizontal(const v3f &target_speed, const f32 max_increase);
-	void accelerateVertical(const v3f &target_speed, const f32 max_increase);
 	bool updateSneakNode(Map *map, const v3f &position, const v3f &sneak_max);
 	float getSlipFactor(Environment *env, const v3f &speedH);
-	void _applyControl(const ControlLogEntry &cle, Environment *env);
 
 	v3f m_position;
 	v3s16 m_standing_node;
@@ -157,8 +131,6 @@ private:
 	v3s16 m_sneak_node = v3s16(32767, 32767, 32767);
 	// Stores the top bounding box of m_sneak_node
 	aabb3f m_sneak_node_bb_top = aabb3f(0.0f, 0.0f, 0.0f, 0.0f, 0.0f, 0.0f);
-	// Whether the player is allowed to sneak
-	bool m_sneak_node_exists = false;
 	// Whether a "sneak ladder" structure is detected at the players pos
 	// see detectSneakLadder() in the .cpp for more info (always false if disabled)
 	bool m_sneak_ladder_detected = false;
@@ -174,10 +146,7 @@ private:
 	std::string m_old_node_below_type = "air";
 	// ***** End of variables for temporary option *****
 
-	bool m_can_jump = false;
 	u16 m_breath = PLAYER_MAX_BREATH_DEFAULT;
-	f32 m_yaw = 0.0f;
-	f32 m_pitch = 0.0f;
 	bool camera_barely_in_ceiling = false;
 	aabb3f m_collisionbox = aabb3f(-BS * 0.30f, 0.0f, -BS * 0.30f, BS * 0.30f,
 			BS * 1.75f, BS * 0.30f);
